BUG: tweaking delete method and getting rid of warning messages.
git-svn-id: http://svn.slicer.org/Slicer3/branches/Slicer-3-6@15651 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/vtkURIHandler.cxx
@@ -103,10 +103,13 @@ void vtkURIHandler::DeleteFileBucket()
     }
 
   // Be tidy. delete the file.
-  bool clean = vtksys::SystemTools::RemoveFile ( this->GetFileBucket() );
-  if (!clean )
+  if ( vtksys::SystemTools::FileExists ( this->GetFileBucket() ) )
     {
-    vtkWarningMacro ( "Unable to clean up temporary download file " << this->GetFileBucket() );
+    bool clean = vtksys::SystemTools::RemoveFile ( this->GetFileBucket() );
+    if (!clean )
+      {
+      vtkWarningMacro ( "Unable to clean up temporary download file " << this->GetFileBucket() );
+      }
     }
 
   // set the temporary staging area to null
